Accorpamento di manage_dataset.py in model.py con successivo refactoring
